+++ src/main/webapp/resources/js/apis/samples/samples.js
@@ -1,0 +1,6 @@
+import axios from "axios";
+
+const URL = `${window.TL.BASE_URL}samples`;
+
+export const getSampleDetails = id =>
+  axios.get(`${URL}?id=${id}`).then(response => response.data);