+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Changes
 * [UI]: Fixed bug preventing associated projects from being loaded into the project samples table. (19.01.3)
 * [Developer]: Moved FastQC results out of database to filesystem for a big reduction in database size and performance.  See https://irida.corefacility.ca/documentation/adminstrator/upgrades/#1904 for more information. 
 * [UI]: Added hard wrap on sample name on sample details page.
+* [UI]: New dedicated cart page.
+* [UI]: Exporting to Galaxy now runs through the new cart interface.  Export to Galaxy through the Project Samples page has been deprecated.
 * [Developer]: Updated Travis CI dist to `xenial`.
 * [Developer]: Updated chromedriver in pom.xml and packages.json to newer versions to work better with newest chrome version.